Add back flake8-bugbear B008 ignore
## `B008`

Added back ignoring of `flake8-bugbear` `B008`.

This rule seems to be targeting functions with mutable return values. While it does have a way of allowing specific functions to get through, the list is not comprehensive and doesn't even cover the default python namespace, let alone the standard library. Examples of things it doesn't catch: `slice`, `Path`, `MappingProxyType`

The error is also bad. It says:

> B008 Do not perform function calls in argument defaults.  The call is performed only once at function definition time. All calls to your function will reuse the result of that definition-time function call.  If this is intended, assign the function call to a module-level variable and use that variable as a default value.

* It misidentifies the problem. The problem is mutable values, not having function calls defining default arguments.
* The suggested solution is working around the ability of flake-8s static analysis to see what has happened. The argument should behave exactly the same afterwards.
* While it's proposed solution is a workaround instead of a fix, it does not document the `extend-immutable-calls`