Fix template sync due to lacking permissions (#137)
In a test-run, the template sync failed since apparently
the `sync.yaml` action is not permitted to make changes to files in the
`.github/workflows` directory.

This is a bit of a problem for which we still need to figure out a
solution. One of them is to use a personal access token. Another one
could be to centralize template sync, which I would have liked to avoid.

For now, I just rolled back the changes to the workflow files since
the last release since they were mostly of cosmetic nature.